Move all code to common base package 'bingo'
