--- src/main/java/achievements/modifiers/PointValueModifier.java
@@ -1,10 +1,0 @@
-package achievements.modifiers;
-
-import ribbons.Ribbon;
-
-public record PointValueModifier(Ribbon ribbon, double bonusModifier) {
-
-    public String getBonusModifierAsPercentage() {
-        return Math.round(bonusModifier * 100) + "%";
-    }
-}
