--- src/main/java/application/Launcher.java
@@ -1,9 +1,0 @@
-package application;
-
-import application.gui.BingoUserInterface;
-
-public class Launcher {
-    public static void main(String[] args) {
-        BingoUserInterface.main(args);
-    }
-}
