--- src/main/java/application/gui/input/UserInputException.java
@@ -1,8 +1,0 @@
-package application.gui.input;
-
-public class UserInputException extends Exception {
-
-    public UserInputException(String message, Exception cause) {
-        super(message, cause);
-    }
-}
