+++ src/main/java/bingo/achievements/modifiers/PointValueModifier.java
@@ -1,0 +1,10 @@
+package bingo.achievements.modifiers;
+
+import bingo.ribbons.Ribbon;
+
+public record PointValueModifier(Ribbon ribbon, double bonusModifier) {
+
+    public String getBonusModifierAsPercentage() {
+        return Math.round(bonusModifier * 100) + "%";
+    }
+}
