+++ src/main/java/bingo/application/Launcher.java
@@ -1,0 +1,9 @@
+package bingo.application;
+
+import bingo.application.gui.BingoUserInterface;
+
+public class Launcher {
+    public static void main(String[] args) {
+        BingoUserInterface.main(args);
+    }
+}
