+++ src/main/java/bingo/application/gui/input/UserInputException.java
@@ -1,0 +1,8 @@
+package bingo.application.gui.input;
+
+public class UserInputException extends Exception {
+
+    public UserInputException(String message, Exception cause) {
+        super(message, cause);
+    }
+}
