+++ src/main/java/bingo/ribbons/WeightedRibbonResult.java
@@ -1,0 +1,18 @@
+package bingo.ribbons;
+
+import bingo.points.WeightedPointResult;
+import bingo.ships.MainArmamentType;
+
+public record WeightedRibbonResult(RibbonResult ribbonResult, MainArmamentType mainArmamentType)
+        implements WeightedPointResult {
+
+    @Override
+    public int getPointValue() {
+        return ribbonResult.getPointValue(mainArmamentType);
+    }
+
+    @Override
+    public String getAsString() {
+        return ribbonResult.getAsString(mainArmamentType);
+    }
+}
