+++ src/main/java/bingo/ribbons/overrides/PointValueOverride.java
@@ -1,0 +1,6 @@
+package bingo.ribbons.overrides;
+
+import bingo.ships.MainArmamentType;
+
+public record PointValueOverride(MainArmamentType mainArmamentType, int pointValue) {
+}
