--- src/main/java/ribbons/WeightedRibbonResult.java
@@ -1,18 +1,0 @@
-package ribbons;
-
-import bingo.points.WeightedPointResult;
-import ships.MainArmamentType;
-
-public record WeightedRibbonResult(RibbonResult ribbonResult, MainArmamentType mainArmamentType)
-        implements WeightedPointResult {
-
-    @Override
-    public int getPointValue() {
-        return ribbonResult.getPointValue(mainArmamentType);
-    }
-
-    @Override
-    public String getAsString() {
-        return ribbonResult.getAsString(mainArmamentType);
-    }
-}
