--- src/main/java/ribbons/overrides/PointValueOverride.java
@@ -1,6 +1,0 @@
-package ribbons.overrides;
-
-import ships.MainArmamentType;
-
-public record PointValueOverride(MainArmamentType mainArmamentType, int pointValue) {
-}
