+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,5 +1,5 @@
 Manifest-Version: 1.0
-Main-Class: application.Launcher
+Main-Class: bingo.application.Launcher
 Class-Path: dependencies/javafx-fxml-23.0.2.jar dependencies/javafx-fxml-23.0.2-win.jar dependencies/javafx-con
  trols-23.0.2-win.jar dependencies/javafx-graphics-23.0.2-win.jar dependencies/javafx-graphics-23.
  0.2.jar dependencies/javafx-controls-23.0.2.jar dependencies/javafx-base-23.0.2.jar dependencies/javafx-base-2
